refactor(math): split quaternion kernel into TrajectoryMath.Quaternions.cs
